requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
daphne==3.0.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap-dynamic-formsets==0.5.0
django-bootstrap-pagination==1.7.1
django-bootstrap3==12.0.3
django-cors-headers==3.5.0
django-crontab==0.7.1
django-localflavor==3.0.1
django-model-utils==4.1.1
django-nocaptcha-recaptcha==0.0.20
django-notifications-hq==1.6.0
django-rest-swagger==2.2.0
django-select2==7.2.3
django-user-agents==0.4.0
djangorestframework==3.11.0
hiredis==1.1.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
itypes==1.1.0
Jinja2==2.11.1
jsonfield==3.1.0
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-social-auth==0.3.6
python-stdnum==1.14
python3-openid==3.1.0
pytz==2019.3
raven==6.10.0
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
swapper==1.1.2.post1
Twisted==20.3.0
txaio==20.12.1
ua-parser==0.10.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.2.0
webencodings==0.5.1
zope.interface==5.2.0
